+++ src/Akeneo/Pim/WorkOrganization/Workflow/Component/Model/PublishedProduct.php
@@ -52,25 +52,11 @@ class PublishedProduct implements ReferableInterface, PublishedProductInterface
     /** @var array */
     protected $rawValues;
 
-    /** @var \Datetime $created */
+    /** @var \DateTime $created */
     protected $created;
 
-    /** @var \Datetime $updated */
+    /** @var \DateTime $updated */
     protected $updated;
-
-    /**
-     * Not persisted but allow to force locale for values
-     *
-     * @var string
-     */
-    protected $locale;
-
-    /**
-     * Not persisted but allow to force scope for values
-     *
-     * @var string
-     */
-    protected $scope;
 
     /**
      * Not persisted. Loaded on the fly via the $rawValues.
@@ -184,42 +170,6 @@ class PublishedProduct implements ReferableInterface, PublishedProductInterface
     public function setUpdated($updated)
     {
         $this->updated = $updated;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getScope()
-    {
-        return $this->scope;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setScope($scope)
-    {
-        $this->scope = $scope;
 
         return $this;
     }